Refactor Promo Card components and update README
- Consolidated card variants into a single Card component for better maintainability.
- Updated README.md to reflect changes in card structure and removed outdated sections.
- Deleted unused SVG files to clean up the public directory.
- Adjusted TypeScript interfaces for improved clarity and consistency in props.
- Removed unused utility function for random station selection in stationUtils.ts.
